added full hex, octal and other bases and pointers support

+++ libft/srcs/ft_atoi_base.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-static int	skip_obs(char *str, int *is_neg)
+static int		skip_obs(char *str, int *is_neg)
 {
 	int		i;
 
@@ -31,12 +31,12 @@ static int	skip_obs(char *str, int *is_neg)
 	return (i);
 }
 
-long		ft_atoi_base(char *str, int base)
+long long		ft_atoi_base(char *str, int base)
 {
-	int		i;
-	int		is_neg;
-	long	res;
-	int		digits;
+	int			i;
+	int			is_neg;
+	long long	res;
+	int			digits;
 
 	res = 0;
 	is_neg = 1;
